bug fixed!
found the reason why ch11 test#7 finding the refracted color failed!
in the function `prepare_computations`, we flip the sign of normalv if
ray origin is inside the object. and the over_point, under_point are
supposed to be calculated after the sign invertion is done.

however, with this bug fixed, the color calculated in test#7 are
slightly deviated from the expected answer from the book. similiar issue
occur at some guy's github,
github.com/toku345/RayTracerChallenge_Python/